+++ config/initializers/shellwords.rb
@@ -1,0 +1,8 @@
+module Shellwords
+  # returns args, pipe
+  def self.parse(expression)
+    tokens, pipe = expression.split('|', 2)
+    pipe = pipe.strip.tr('‘', "'").tr('“', '"') if pipe
+    [Shellwords.shellwords(tokens), pipe]
+  end
+end