+++ slack-metabot/commands/slack.rb
@@ -9,10 +9,18 @@ module SlackMetabot
         expression = match['expression']
         expression.gsub! '—', '--'
         logger.info "SLACK: #{client.team} - #{expression}"
-        args = Shellwords.shellwords(expression)
-        execute(client, args) do |output|
-          send_message client, data.channel, "```\n#{output}```"
+        args, pipe = Shellwords.parse(expression)
+        execute(client, args) do |output, error|
+          if error
+            send_message client, data.channel, "```\n#{error}```"
+          else
+            output = pipe ? JsonPath.on(output, pipe) : JSON.parse(output)
+            output = JSON.pretty_generate(output)
+            send_message client, data.channel, "```\n#{output}```"
+          end
         end
+      rescue SyntaxError => e
+        send_message_with_gif client, data.channel, e.message, 'error'
       end
 
       def self.execute(client, args)
@@ -22,8 +30,7 @@ module SlackMetabot
           client.team.token,
           args
         ].flatten) do |_, stdout, stderr, _|
-          output = stderr.gets || stdout.gets
-          yield output.strip
+          yield stdout.gets.try(:strip), stderr.gets.try(:strip)
         end
       end
     end