+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ### Changelog
 
-* 1/5/2016: Initial public release - [@dblock](https://github.com/dblock).
+### 1/6/2016
+
+* [#1](https://github.com/dblock/slack-metabot/issues/1): Added JSON processing - [@dblock](https://github.com/dblock).
+* JSON output is now prettier - [@dblock](https://github.com/dblock).
+
+### 1/5/2016
+
+* Initial public release - [@dblock](https://github.com/dblock).